Added confirmation box for delete

+++ src/views/Challenges/Setup/SetupAbstract.vue
@@ -132,13 +132,40 @@ export default class Setup extends Vue {
 
   async deleteG() {
     const goal = this.goal(this.goal_url);
+    const self = this;
 
-    if (goal._links.edit) {
-      await this.deleteGoal(goal._links.edit.href);
-      this.$router.push({
-        name: "ChallengesOverview",
-        params: { type: this.type }
-      });
+    //@ts-ignore
+    if (navigator && navigator.notification) {
+      //@ts-ignore
+      navigator.notification.confirm(
+        self.$t("challengedeletemsg"),
+        async (result: number) => {
+          if (result === 1) {
+            //Cancel
+          } else if (result === 2) {
+            if (goal._links.edit) {
+              await this.deleteGoal(goal._links.edit.href);
+              this.$router.push({
+                name: "ChallengesOverview",
+                params: { type: this.type }
+              });
+            }
+          }
+        },
+        '',
+        [self.$t("No"), self.$t("Yes")]
+      );
+    } else {
+      let confirm_response: any = confirm((self.$t("challengedeletemsg") as string));
+      if(confirm_response){
+        if (goal._links.edit) {
+          await this.deleteGoal(goal._links.edit.href);
+          this.$router.push({
+            name: "ChallengesOverview",
+            params: { type: this.type }
+          });
+        }
+      }
     }
   }
 }
